Refactor: move mypage into member
