+++ src/main/java/com/manittotie/manilib/member/service/MemberService.java
@@ -4,7 +4,10 @@ import com.manittotie.manilib.member.domain.Member;
 import com.manittotie.manilib.member.dto.DuplicateEmailRequest;
 import com.manittotie.manilib.member.dto.DuplicateEmailResponse;
 import com.manittotie.manilib.member.dto.MemberDto;
+import com.manittotie.manilib.member.dto.MessageRequest;
 import com.manittotie.manilib.member.repository.MemberRepository;
+import com.manittotie.manilib.member.dto.MessageResponse;
+import com.manittotie.manilib.member.dto.MyPageResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -36,6 +39,34 @@ public class MemberService {
         return memberRepository.findAllByGroupId(groupId).stream()
                 .map(m-> new MemberDto(m.getId(), m.getEmail(), m.getNickname()))
                 .collect(Collectors.toList());
+    }
+
+    // 마이페이지 조회
+    public MyPageResponse GetMyPage(Long memberId) {
+        Member member = memberRepository.findById(memberId)
+                .orElseThrow(() -> new IllegalArgumentException("해당 맴버가 존재하지 않습니다"));
+
+        MyPageResponse response = new MyPageResponse();
+
+        response.setNickname(member.getNickname());
+        response.setEmail(member.getEmail());
+        response.setPassword(member.getPassword());
+        response.setMyMessage(member.getMyMessage());
+
+        return response;
+    }
+
+    // 상태메세지 작성
+    public MessageResponse WriteMyMessage(MessageRequest request, String email) {
+        Member member = memberRepository.findByEmail(email)
+                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 사용자입니다"));
+
+        member.setMyMessage(request.getMessage());
+
+        MessageResponse response = new MessageResponse();
+        response.setMessage(request.getMessage());
+
+        return response;
     }
 
 }
